requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
passlib[bcrypt]
python-jose
motor
